Separo el uploaded file en su propio archivo
